refactor: deduplicate menu drawing helpers into menu_common
Move draw_help, draw_status_message, confirm_dialog, and
clear_line_remainder into menu_common.rs as shared implementations.
Replace per-menu duplicates in menu.rs, cfr_menu.rs, and
secure_boot_menu.rs with thin delegating wrappers. Also remove
trivial read_key/clear_screen forwarding wrappers.
